requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.2
google-cloud-secret-manager==2.20.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
gunicorn==22.0.0
idna==3.7
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
openai==1.3.7
openpyxl==3.1.4
orjson==3.10.3
pandas==2.2.2
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
ruff==0.4.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.31
starlette==0.37.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
